new script for translating words
tw.pl translates words using WordReference or Wikipedia. The languages
default to "English -> French", but they can be changed with command line
options -f and -t. Using Wikipedia is useful for translating technical
terms that have no entry in the dictionary but that have articles in the
desired language on Wikipedia.